+++ modules/graphics/tests/test_contour3.m
@@ -14,9 +14,9 @@ x = linspace(-2 * pi, 2 * pi);
 y = linspace(0, 4 * pi);
 [X, Y] = meshgrid(x, y);
 Z = sin(X) + cos(Y);
-contour(X, Y, Z);
+contour3(X, Y, Z);
 %=============================================================================
-[M, h] = contour(X, Y, Z);
+[M, h] = contour3(X, Y, Z);
 R = properties(h);
 REF = {'ContourMatrix';
 'Children';
@@ -42,40 +42,40 @@ REF = {'ContourMatrix';
 'ZData'};
 assert_isequal(R, REF);
 assert_isequal(M, h.ContourMatrix);
-assert_isequal(h.Floating, 'off');
+assert_isequal(h.Floating, 'on');
 %=============================================================================
 figure();
 [X, Y, Z] = peaks;
-contour(X, Y, Z, 20)
+contour3(X, Y, Z, 20)
 %=============================================================================
 figure();
 [X, Y, Z] = peaks;
 v = [1,1];
-contour(X, Y, Z, v)
+contour3(X, Y, Z, v)
 %=============================================================================
 figure();
 [X, Y, Z] = peaks;
-contour(X, Y, Z, '-.')
+contour3(X, Y, Z, '-.')
 %=============================================================================
 figure();
 Z = peaks;
-[M, c] = contour(Z);
+[M, c] = contour3(Z);
 c.LineWidth = 3;
 %=============================================================================
 figure();
 Z = peaks;
 Z(:,26) = NaN;
-contour(Z)
+contour3(Z)
 %=============================================================================
 figure();
 [theta, r] = meshgrid (linspace (0,2*pi,64), linspace (0,1,64));
 [X, Y] = pol2cart (theta, r);
 Z = sin (2*theta) .* (1-r);
-contour (X, Y, abs (Z), 10);
+contour3(X, Y, abs (Z), 10);
 %=============================================================================
 figure();
 [x, y, z] = peaks ();
-[c, h] = contour (x, y, z);
+[c, h] = contour3 (x, y, z);
 assert_isequal(h.LevelStepMode, 'auto');
 h.LevelStep = 3;
 assert_isequal(h.LevelStepMode, 'manual');