+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `contour` Contour plot of matrix.
+- `contour3` 3-D contour plot.
 
 ### Fixed